requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

PyPDF2==3.0.1
langchain==0.1.4
langchain-text-splitters==0.0.1

sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

faiss-cpu==1.7.4
numpy==1.26.3

google-generativeai==0.3.2

python-dotenv==1.0.0